+++ framework/base/Model.php
@@ -575,6 +575,27 @@ class Model extends Component implements IteratorAggregate, ArrayAccess, Arrayab
     }
 
     /**
+     * Adds a list of errors.
+     * @param array $errors a list of errors. The array keys must be attribute names.
+     * The array values should be error messages. If an attribute has multiple errors,
+     * these errors must be given in terms of an array.
+     * You may use the result of [[getErrors()]] as the value for this parameter.
+     * @since 2.0.2
+     */
+    public function addErrors(array $errors)
+    {
+        foreach ($errors as $attribute => $error) {
+            if (is_array($error)) {
+                foreach($error as $e) {
+                    $this->addError($attribute, $e);
+                }
+            } else {
+                $this->addError($attribute, $error);
+            }
+        }
+    }
+
+    /**
      * Removes errors for all attributes or a single attribute.
      * @param string $attribute attribute name. Use null to remove errors for all attribute.
      */